+++ Code/Tools/GuiFoundation/Widgets/ColorGradientEditorWidget.moc.h
@@ -25,18 +25,18 @@ public:
   void FrameGradient();
 
 signals:
-  void ColorCpAdded(float posX, const ezColorGammaUB& color);
+  void ColorCpAdded(double posX, const ezColorGammaUB& color);
   void ColorCpMoved(ezInt32 index, float newPosX);
   void ColorCpDeleted(ezInt32 index);
   void ColorCpChanged(ezInt32 index, const ezColorGammaUB& color);
 
-  void AlphaCpAdded(float posX, ezUInt8 alpha);
-  void AlphaCpMoved(ezInt32 index, float newPosX);
+  void AlphaCpAdded(double posX, ezUInt8 alpha);
+  void AlphaCpMoved(ezInt32 index, double newPosX);
   void AlphaCpDeleted(ezInt32 index);
   void AlphaCpChanged(ezInt32 index, ezUInt8 alpha);
 
-  void IntensityCpAdded(float posX, float intensity);
-  void IntensityCpMoved(ezInt32 index, float newPosX);
+  void IntensityCpAdded(double posX, float intensity);
+  void IntensityCpMoved(ezInt32 index, double newPosX);
   void IntensityCpDeleted(ezInt32 index);
   void IntensityCpChanged(ezInt32 index, float intensity);
